+++ mona_codegen/src/lib.rs
@@ -1,3 +1,3 @@
 //! Mona crate responsible for the generation of the LLVM IR.
 
-pub mod complier;
+pub mod gen;
