+++ mona_compiler/src/lib.rs
@@ -1,3 +1,4 @@
 //! This is the compiler of Mona
 //!
 //! This file is responsible for the transformation of the LLVM IR to object files and binary / lib.
+
